feat(API): Exception types for the API to report.
The `_libs` exceptions for dealing with `ctypes` are being updated to use
the new API exceptions.  This deviates from `nidaqmx` but is valuable for
allowing a client to catch `ctypes` related exceptions.  This also means
that the `errors` module will need to be structured differently to avoid
circular imports.

Also, this deviates from DAQmx in that warnings do not have a preceding
newline.  That is being left for the user to decide how they want it
formatted.