feat(API): Implement read_signal_single_point
Unfortunately, there isn't a way to ask nxReadSignalSinglePoint what size
of buffer we should create, so we have to take in how many signals exist.
This suggests that we'll want the OOP API to have readers/writers like
DAQmx so we can ask the session for the number of signals once and reuse
it across calls.