Minor API changes to get the signal single point example working
Change nxStatus_t to be a u32 instead of i32. The problem is our constants are defined as unsigned while nxStatus_t was defined as signed. This caused several problems (incorrect detection of errors, failure to convert code to enum value).
For now, we are going with the "quickest to get working" solution. In the long term, we need to decide if we should treat these as unsigned or signed.