feat(API): Move interface/j1939 session properties
These are moved to an `Interface` and `J1939` child objects.

`Interface` is for properties related to the device's interface and makes sense as a
standalone object, similar to how frames will be standalone objects.

`J1939` is less clear.  This might be viewed more as having a `Timing`
object as part of a DAQmx task.

Additional benefits
- Fewer properties to overload a developer when their IDE offers them
  auto-complete options.
- Consistent with the plan for frame properties though those you have to
  select from.
- Easier to reuse properties across multiple Session classes if we go that
  route.