feat(tests): Example of integration tests
This is leveraging py.test for both unit and integration tests.

Hardware is auto-selected but can be overridden with an environment
variable.  A `fixture` was used to pass this to the test (rather than the
test reading from `request` directly) so that if the test fails, pytest
will show what hardware fixture was used.

We should not run these integration tests as part of TravisCI though.  To
accomodate this, a distinct tox config is used for integration tests. It
needs to list all of the same environments as the main one because
`tox-travis` reads it for the possible environments and we want to be able
to verify the syntax of the file.

A single tox config was attempted but there didn't seem to be a good way
to conditionally only run part of the environment.  Using `TOXENV` was
attempted but that didn't play well with `tox-travis`.

Another approach could be to reduce the value of `tox-travis` and
explicitly declare each part of the test matrix to ensure any
`py*-integration` tests aren't run.