fix(API): Define equality semantics
In finishing the major container semantics for `Frames` and `Signals, a
decision neede to be made for them and `Interface` on what should be
compared for equality. There are two main points of comparison these
sub-objects, associated resource and configuration.

For configuration, that is effectively comparing `(handle, index)` which
doesn't have too much value.

For associated resource, the comparison is `(db_name, item_name)`.  This
might have some use, especially if it later integrates with `System.intfs`
and database `Frames` and `Signals`.  The problem is it adds complexity
and a performance hit.  For now, we are only applying this to
`Session.intf` since it doesn't have a db_name to worry about.