fix(CI): Only run tests under test environments
So the hacks pile up.  We have a separate file to avoid running the tests
on TravisCI.  We need all the main environments listed in this separate
file so we can do syntax checks on it (because of how tox-travis works).
The problem is that then caused our tests to be run under all of those
environments on Jenkins but the dependencies weren't installed.

So limiting the test run to the test environment seems to make everything
happy.