feat(API): Provide can ID parser for `out_strm_list_by_id`
`out_strm_list_by_id` might contain either LIN or CAN frames IDs without a
way to distinguish them so we can't completely abstract it like we did for
IDs in the actual Frames.  Instead the parsing logic from `CanFrame` was
extracted and abstracted so it can be reused with this property
(`CanIdentifier`).

Fixes #115

BREAKING CHANGE: `CanFrame` now accepts a single `CanIdentifier` rather
than an `(int, bool)`.  An `int` can still be passed in as a short-form
and will be interpreted as a non-extended id.