fix: Improve thread safety
While nixnet doesn't have callbacks to exacerbatge the problem, still
wanting to protect things.  nidaqmx went with a per-function lock.  That
is probably overkill for our needs so going with a global library loading
lock.

Not introducing double-check locking into the picture because the
conditional initialization is just an optimization.  nixnet should not
have any problems if two threads make it into the initialization check,
especially considering every call did initalization in the past.

Fixes #146