feat(API): LinBusErrorFrame support
Renamed `LinComm` attributes from previous commit to align with names from
`LinBusErrorFrame` in a similar way to how `CanComm` alignswith
`CanBusErrorFrame`.  This still leaves `last_err` vs `bus_err`.  Unsure
if it should change and didn't want to break it in this build.

BREAKING CHANGE: Renamed `LinLastErrCode` to `LinLastErr` to align with
`CanLastErr`.