fix(API): Pass null-terminated char* to driver
This fixes an Ascii decoding bug in the following three properties:

* DatabaseObject.comment
* Frames.set_session_j1939_addr_filter
* Interface.src_term_start_trigger

When using these property setters, we were passing
a char* without null termination to the driver.
Then, when reading these properties with the
getter, we'd get our string back with non-Ascii
characters on the end, which caused an exception
when trying to decode to Ascii.

In the cprops layer for these properties,
we converted strings to a char* using
create_string_buffer with a length argument,
which results in copying over the python
string into the buffer without a null terminator.

Other places in the _funcs.py would just convert
the python strings to a char_p before passing to
the driver, and doing this doesn't guarantee a
null-terminted char*.

The best strategy for converting python strings
to a null-terminated char* is to use
create_string_buffer without a length argument.